+++ src/main/java/um/tesoreria/core/hexagonal/articulo/infrastructure/web/controller/ArticuloController.java
@@ -1,11 +1,11 @@
 /**
  * 
  */
-package um.tesoreria.core.controller;
+package um.tesoreria.core.hexagonal.articulo.infrastructure.web.controller;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,9 +18,9 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import um.tesoreria.core.exception.ArticuloException;
-import um.tesoreria.core.kotlin.model.Articulo;
+import um.tesoreria.core.hexagonal.articulo.infrastructure.persistence.entity.ArticuloEntity;
 import um.tesoreria.core.model.view.ArticuloKey;
-import um.tesoreria.core.service.ArticuloService;
+import um.tesoreria.core.hexagonal.articulo.application.service.ArticuloService;
 
 /**
  * @author daniel
@@ -28,47 +28,47 @@ import um.tesoreria.core.service.ArticuloService;
  */
 @RestController
 @RequestMapping("/articulo")
+@RequiredArgsConstructor
 public class ArticuloController {
 
-	@Autowired
-	private ArticuloService service;
+	private final ArticuloService service;
 
 	@GetMapping("/")
-	public ResponseEntity<List<Articulo>> findAll() {
-		return new ResponseEntity<List<Articulo>>(service.findAll(), HttpStatus.OK);
+	public ResponseEntity<List<ArticuloEntity>> findAll() {
+		return new ResponseEntity<>(service.findAll(), HttpStatus.OK);
 	}
 
 	@PostMapping("/search")
 	public ResponseEntity<List<ArticuloKey>> findByStrings(@RequestBody List<String> conditions) {
-		return new ResponseEntity<List<ArticuloKey>>(service.findByStrings(conditions), HttpStatus.OK);
+		return new ResponseEntity<>(service.findByStrings(conditions), HttpStatus.OK);
 	}
 
 	@GetMapping("/{articuloId}")
-	public ResponseEntity<Articulo> findByArticuloId(@PathVariable Long articuloId) {
+	public ResponseEntity<ArticuloEntity> findByArticuloId(@PathVariable Long articuloId) {
 		try {
-			return new ResponseEntity<Articulo>(service.findByArticuloId(articuloId), HttpStatus.OK);
+			return new ResponseEntity<>(service.findByArticuloId(articuloId), HttpStatus.OK);
 		} catch (ArticuloException e) {
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
 		}
 	}
 
 	@GetMapping("/new")
-	public ResponseEntity<Articulo> articuloNew() {
+	public ResponseEntity<ArticuloEntity> articuloNew() {
 		try {
-			return new ResponseEntity<Articulo>(service.articuloNew(), HttpStatus.OK);
+			return new ResponseEntity<>(service.articuloNew(), HttpStatus.OK);
 		} catch (ArticuloException e) {
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
 		}
 	}
 
 	@PostMapping("/")
-	public ResponseEntity<Articulo> add(@RequestBody Articulo articulo) {
-		return new ResponseEntity<Articulo>(service.add(articulo), HttpStatus.OK);
+	public ResponseEntity<ArticuloEntity> add(@RequestBody ArticuloEntity articulo) {
+		return new ResponseEntity<>(service.add(articulo), HttpStatus.OK);
 	}
 	
 	@PutMapping("/{articuloId}")
-	public ResponseEntity<Articulo> update(@RequestBody Articulo articulo, @PathVariable Long articuloId) {
-		return new ResponseEntity<Articulo>(service.update(articulo, articuloId), HttpStatus.OK);
+	public ResponseEntity<ArticuloEntity> update(@RequestBody ArticuloEntity articulo, @PathVariable Long articuloId) {
+		return new ResponseEntity<>(service.update(articulo, articuloId), HttpStatus.OK);
 	}
 
 }