+++ src/main/java/um/tesoreria/core/hexagonal/auth/infrastructure/web/controller/AuthController.java
@@ -32,7 +32,7 @@ public class AuthController {
         } catch (IllegalArgumentException e) {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
         } catch (Exception e) {
-            throw new  ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
         }
     }
 }