+++ src/main/java/um/tesoreria/core/model/dto/CostoParameterDto.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import um.tesoreria.core.kotlin.model.Comprobante;
 import um.tesoreria.core.kotlin.model.Ubicacion;
-import um.tesoreria.core.kotlin.model.Articulo;
+import um.tesoreria.core.hexagonal.articulo.infrastructure.persistence.entity.ArticuloEntity;
 import um.tesoreria.core.model.UbicacionArticulo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -25,7 +25,7 @@ public class CostoParameterDto implements Serializable {
 
 	private static final long serialVersionUID = 3389089202450536124L;
 
-	private List<Articulo> articulos;
+	private List<ArticuloEntity> articulos;
 	private List<Comprobante> comprobantes;
 	private List<Ubicacion> ubicacions;
 	private List<UbicacionArticulo> ubicacionArticulos;