+++ src/main/java/um/tesoreria/core/service/dto/CostoParameterService.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import um.tesoreria.core.model.dto.CostoParameterDto;
-import um.tesoreria.core.service.ArticuloService;
+import um.tesoreria.core.hexagonal.articulo.application.service.ArticuloService;
 import um.tesoreria.core.service.ComprobanteService;
 import um.tesoreria.core.service.UbicacionArticuloService;
 import um.tesoreria.core.service.UbicacionService;